検索初期表示まで

+++ src/model/ddl/common.go
@@ -9,9 +9,9 @@ type Notice struct {
 	// 種別
 	Type uint `json:"type"`
 	// 通知元ユーザーID
-	FromUserID uint `json:"from_user_id"`
+	FromUserID uint64 `json:"from_user_id"`
 	// 通知先ユーザーID
-	ToUserID uint `json:"to_user_id"`
+	ToUserID uint64 `json:"to_user_id"`
 	// 通知種別(外部キー)
 	NoticeType NoticeType `gorm:"foreignKey:type;references:id"`
 	// 通知元ユーザー(外部キー)
@@ -24,10 +24,10 @@ func (t Notice) TableName() string {
 	return "t_notice"
 }
 
-// 共通Entity
+// 共通Entity ※削除予定
 type CommonModel struct {
 	// ID
-	ID uint `json:"id"`
+	ID uint64 `json:"id"`
 	// ハッシュキー
 	HashKey uint `json:"hash_key"`
 }
