+++ src/model/response/error.go
@@ -4,14 +4,14 @@ type Error struct {
 	// スタータス(ヘッダー)
 	Status int `json:"status"`
 	// コード
-	Code int8 `json:"code"` // 必要な場合のみ
+	Code uint `json:"code"` // 必要な場合のみ
 	// エラー
 	Error error `json:"error"` // 消す可能性大
 }
 
 type ErrorCode struct {
 	// コード
-	Code int8 `json:"code"`
+	Code uint `json:"code"`
 }
 
 func ErrorConvert(e Error) ErrorCode {
